SoulSyncApp/infra/monitoring/therapy-engine-release-diamond-k8s-aiops-enterprise/therapy-engine-release-diamond-k8s-aiops-enterprise/therapy-engine-release-diamond-k8s-aiops-guardrails/therapy-engine-release-diamond-k8s-aiops/therapy-engine-release-diamond-k8s-kpi/therapy-engine-release-diamond-k8s-slo/therapy-engine-release-diamond-k8s-budget/therapy-engine-release-diamond-k8s-cost/therapy-engine-release-diamond-k8s-security/therapy-engine-release-diamond-k8s-log-alerts/therapy-engine-release-diamond-k8s-logging/therapy-engine-release-diamond-k8s-tracing/therapy-engine-release-diamond-k8s-cicd/therapy-engine-release-diamond-k8s-observability/therapy-engine-release-diamond-k8s-grafana/therapy-engine-release-diamond-k8s-alerting/therapy-engine-release-diamond-k8s-monitoring/therapy-engine-release-diamond-k8s-alerts/therapy-engine-release-diamond-k8s-notify/therapy-engine-release-diamond-k8s-drill/therapy-engine-release-diamond-k8s-failover/therapy-engine-release-diamond-k8s-dr/therapy-engine-release-diamond-k8s-snapshots/therapy-engine-release-diamond-k8s-backup/therapy-engine-release-diamond-k8s-multicloud/therapy-engine-release-diamond-k8s-azure/therapy-engine-release-diamond-k8s-gcp/therapy-engine-release-diamond-k8s-db/therapy-engine-release-diamond-k8s-tf/therapy-engine-release-diamond-k8s-ci/therapy-engine-release-diamond-scale/therapy-engine-release-diamond-multidomain/therapy-engine-release-diamond-ssl/therapy-engine-release-diamond-deploy/therapy-engine-release-diamond-ci/therapy-engine-release-diamond-plus/therapy-engine-release-diamond/ops/aiops/detector/requirements.txt
requests
numpy
prometheus_client
scikit-learn==1.5.1

psycopg2-binary==2.9.9
